moved utils.py to cmd_utils.py to contain cli utils so that utils.py can contain general ones.
